requirements.txt
azure-graphrbac
pyyaml